requirements.txt
python-dotenv==1.0.1
requests==2.32.4
google-api-python-client==2.172.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0